uadk: adjust header file include logic
Adjust the header file inclusion logic to ensure that the data
structures used internally are not exposed to the outside world.

Clarify the file relationship diagram of UADK, follow-up UADK
framework evolution, continue to ensure that the logic of the
basic file inclusion relationship is reasonable.
