+++ include/drv/wd_comp_drv.h
@@ -8,7 +8,7 @@
 #include <asm/types.h>
 
 #include "../wd_comp.h"
-#include "wd_util.h"
+#include "../wd_util.h"
 
 #ifdef __cplusplus
 extern "C" {
