+++ include/drv/wd_ecc_drv.h
@@ -8,7 +8,7 @@
 #include <asm/types.h>
 
 #include "../wd_ecc.h"
-#include "wd_util.h"
+#include "../wd_util.h"
 
 #ifdef __cplusplus
 extern "C" {
