+++ include/drv/wd_rsa_drv.h
@@ -6,7 +6,7 @@
 #include <asm/types.h>
 
 #include "../wd_rsa.h"
-#include "wd_util.h"
+#include "../wd_util.h"
 
 #ifdef __cplusplus
 extern "C" {
