+++ include/wd_dh.h
@@ -15,10 +15,6 @@
 extern "C" {
 #endif
 
-#define BYTE_BITS			8
-#define BYTE_BITS_SHIFT			3
-#define GET_NEGATIVE(val)		(0 - (val))
-
 typedef void (*wd_dh_cb_t)(void *cb_param);
 
 enum wd_dh_op_type {
